Fix HTTP response header corruption issue
Previously in serve_static(), strings were appended to the `header`
buffer like this:

sprintf(header, "%sConnection: keep-alive\r\n", header);

This usage causes undefined behavior because `header` is used both
as the destination and source in sprintf.

According to C99 7.19.6.6 The sprintf function:
If copying takes place between objects that overlap,
the behavior is undefined.

In practice, this corrupted the HTTP headers, causing clients
(such as curl) to misinterpret the response as HTTP/0.9.

Since HTTP/0.9 does not include status lines and headers,
modern clients (like curl or browsers) treat this as an error or
reject the response outright:

curl: (1) Received HTTP/0.9 when not allowed

This fix uses an offset to append to the header buffer,
avoiding overlapping memory operations and producing a correct
HTTP/1.1 response.